finance: extract pure GL mapping into gl-posting-map.ts
Fixes CI: typecheck (companyFromAccountName now accepts nullable
accountName) and test (the mapping/helpers live in an IO-free module so
the unit test no longer imports prisma/supabase at load).
